Don't accept bad input to Tile constructor
An exception *should* be thrown if the value passed for friction is not
a float, because that is an error. Any code that constructs a Tile
instance should have a correct friction value. If the friction value is
coming from user input (e.g., a tiledef file), then the code parsing the
file had better sanitize the input before calling Tile().

Also:
* Fix the module doctest so it actually passes.

To-do:
* Write more doctests.